Added CLI documentation. Moved CLI tests to correct class.
